pyproject.toml
[project]
name = "fastapi-sync-async-starter"
version = "1.1.0"
description = "FastAPI project with sync and async support"
authors = [
    {name = "ysskrishna",email = "sivasaikrishnassk@gmail.com"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12,<4.0"
keywords = [
    "fastapi",
    "async",
    "sync",
    "api",
    "starter",
    "template",
    "docker",
    "web-framework",
    "rest-api",
    "http",
    "postgresql",
    "sqlalchemy",
    "asyncpg",
    "database",
    "boilerplate",
    "starter-kit",
    "production-ready",
    "development",
    "async-support",
    "pytest",
    "testing",
    "coverage",
    "containerization",
    "docker-compose",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

[project.urls]
Homepage = "https://github.com/ysskrishna/fastapi-sync-async-starter"
Documentation = "https://github.com/ysskrishna/fastapi-sync-async-starter#readme"
Repository = "https://github.com/ysskrishna/fastapi-sync-async-starter.git"
"Bug Tracker" = "https://github.com/ysskrishna/fastapi-sync-async-starter/issues"

[tool.poetry.dependencies]
annotated-types = ">=0.7.0"
anyio = ">=4.8.0"
asyncpg = ">=0.29.0"
certifi = ">=2025.1.31"
click = ">=8.1.8"
colorama = ">=0.4.6"
dnspython = ">=2.7.0"
email-validator = ">=2.2.0"
fastapi = ">=0.109.2"
greenlet = ">=3.1.1"
h11 = ">=0.14.0"
httpcore = "==1.0.7"
httpx = "==0.26.0"
idna = ">=3.10"
jinja2 = ">=3.1.6"
markupsafe = ">=3.0.2"
packaging = ">=24.2"
pluggy = ">=1.5.0"
psycopg2-binary = ">=2.9.9"
pydantic = ">=2.10.6"
pydantic-core = ">=2.27.2"
python-dotenv = ">=1.0.1"
sniffio = ">=1.3.1"
sqlalchemy = ">=2.0.27"
starlette = ">=0.36.3"
typing-extensions = ">=4.12.2"
uvicorn = ">=0.27.1"
slowapi = "^0.1.9"
redis = "^5.2.1"

[tool.poetry.group.dev.dependencies]
coverage = ">=7.6.12"
pytest = ">=8.0.1"
pytest-asyncio = ">=0.23.5"
pytest-cov = ">=6.0.0"
pytest-html = ">=4.1.1"
pytest-metadata = ">=3.1.1"
iniconfig = ">=2.0.0"


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
